refactor: split binding.ts into binding/ folder
