+++ packages/express-cargo/src/binding/index.ts
@@ -1,0 +1,1 @@
+export { bindingCargo, getCargo } from './middleware'
